+++ vidya/__init__.py
@@ -2,5 +2,5 @@ import pathlib
 
 ROOT = pathlib.Path(__file__).parent.parent
 
-__version__ = '0.1.4'
-__update_date__ = '2025-06-18_18:22:09'
+__version__ = '0.1.5'
+__update_date__ = '2025-06-19_18:59:27'